+++ jest-setup.ts
@@ -1,4 +1,4 @@
-import './test-utils/to-have-pattern';
+import './test-utils/to-equal-regex';
 import './test-utils/to-match-groups';
 import './test-utils/to-match-all-groups';
 import './test-utils/to-match-string';
